feat: move enabled mod pak files
